Updated so that create is now create-project
When I tried to run "create test-project" I got an error "Unable to locate an app.js in this directory." I check the apostrophe help and create wasn't a command, but create-project was. I tried "create-project" and it worked, so it looks like that command was updated but this doc wasn't.