Explain how to make use of the apos.images.srcset helper
This PR was opened as a complement to https://github.com/apostrophecms/apostrophe/pull/805.
But since the outcome of that PR was not to add a srcset feature
directly to apostrophe-images-widgets (but rather to introduce just
a template helper `apos.images.srcset`), this change adds a section
to the "Responsive images" how-to about how to write a simple widget
that renders an `<img>` tag with a srcset attribute.