Make layout of documentation landing page more compact
The landing page of the documentation is very frequently visited if
you're using the project to any greater extent. This page has two
issues IMO. Firstly, it's the fact that the cards aren't truly links.
That means if you Cmd/Ctrl+click them, they won't open in a new tab
(oh, the frustration).

Secondly, it's the fact that the cards are so high, you neglect the
sections further down on the page.

I attempted to address these nuisances by making the .card's into actual
links, by removing the .card-link's nexted in them and by being just
a tad more conservative about vertical spacing in the .card's.

This is not an attempt at starting some design by committee process
by opening a PR first and not an issue, but rather I thought it might
be more efficient to open a PR directly, since I would have implemented
by proposed design in the browser anyway.