Added a module for dynamically fetching metadata tags in <head> from apostrophecms.org
This site is hosted in a subcatalog of apostrophecms.org, but it's
totally separate from that site at a code level. This means that changes
to the content of apostrophecms.org won't automatically be reflected on
this site. This isn't really a big issue, but in practice it means that
the favicon on apostrophecms.org/docs is broken, and that the og:image
isn't the same. What I've done here is to add a small module that performs
a dead simple HTTP request to apostrophecms.org, parses the HTML response,
extracts the relevant <link> and <meta> tags, and includes them in the
build of this site. Simple but effective!